+++ CHANGELOG.md
@@ -4,6 +4,17 @@
 >
 > The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### 1.4.1
+
+#### Added
+
+- Add support for randomized speech bubbles.
+- Add "Swim" anim clip key for when the player is swimming.
+
+#### Fixed
+
+- Companions will duplicate when farmhand is exiting an event.
+
 ### 1.4.0
 
 #### Added